Update backend
Strutctured the backend code depending on functionality in:
config
controllers
middlewares
routes
services
utils
