+++ backend/api/utils/fetchUtils.js
@@ -1,0 +1,26 @@
+const fetch = (...args) =>
+  import("node-fetch").then(({ default: fetch }) => fetch(...args));
+
+/**
+ * Enhanced fetch with:
+ * - Automatic retries on failure
+ * - Error handling
+ */
+
+const fetchWithRetry = async (url, options = {}, retries = 3) => {
+  for (let i = 0; i < retries; i++) {
+    try {
+      const response = await fetch(url, options);
+      if (response.ok) return response;
+      throw new Error(`Status: ${response.status}`);
+    } catch (error) {
+      if (i === retries - 1) throw error;
+      await new Promise((resolve) => setTimeout(resolve, 1000 * (i + 1)));
+    }
+  }
+};
+
+module.exports = {
+  fetch,
+  fetchWithRetry,
+};
